% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_excel_date}
\alias{as_excel_date}
\title{Excel date}
\usage{
as_excel_date(x)
}
\arguments{
\item{x}{a numeric. Vector of excel dates}
}
\description{
Function to convert Dates imported from Excel to proper R dates
}
\examples{
as_excel_date(46789)
}
