% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category-row-to-column.R
\name{category-row-to-column}
\alias{category-row-to-column}
\alias{category_row_na}
\alias{category_row_regex}
\alias{category_row_to_column}
\title{Category to Column}
\usage{
category_row_na(x, .col, ..., name = "name")

category_row_regex(x, .col, pattern, name = "name")

category_row_to_column(x, .col, ..., name = "name")
}
\arguments{
\item{x}{a dataframe}

\item{.col}{the column containing category names}

\item{...}{the columns to check for NA}

\item{name}{a character. Name of column}

\item{pattern}{a regex pattern to match category rows}
}
\description{
Category to Column
}
\examples{
d <- data.frame(
  x = c(NA, "A", "B", NA, "A", "B"),
  y = c("a", 1:2, "b", 1:2),
  z = c(NA, "A", "B", NA, "A", "B")
)
category_row_na(d, y, x, z)

category_row_na(d, y, x, z, name = "small")

# Cleaning a Table with two different kinds of category rows
d <- data.frame(
  x = c(NA, NA, "A", "B", NA, NA, "A", "B"),
  y = c("Table a", "Male",  1:2, "Table b", "Female",  1:2),
  z = c(NA, NA, "A", "B", NA, NA, "A", "B")
)

d |>
  category_row_regex(y, "^Table", name = "table") |>
  category_row_na(y, x, z, name = "sex")
}
