% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category-to-column.R
\name{category_row_to_column}
\alias{category_row_to_column}
\title{Category to Column}
\usage{
category_row_to_column(x, val, ..., name = "name")
}
\arguments{
\item{x}{a dataframe}

\item{val}{the column containing category names}

\item{...}{the columns to check for NA}

\item{name}{a character. Name of column}
}
\description{
Category to Column
}
\examples{
d <- data.frame(
  x = c(NA, "A", "B", NA, "A", "B"),
  y = c("a", 1:2, "b", 1:2),
  z = c(NA, "A", "B", NA, "A", "B")
)
category_to_column(d, y, x, z)

category_to_column(d, y, x, z, name = "small")
}
