% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse-row.R
\name{collapse_row}
\alias{collapse_row}
\title{Collapse rows}
\usage{
collapse_row(x, rows, sep = "_", na.rm = TRUE)
}
\arguments{
\item{x}{a dataframe}

\item{rows}{rows to collapse}

\item{sep}{separator. Defaults to "_".}

\item{na.rm}{Should NAs be removed before concatenating?}
}
\description{
Collapse rows into one by concatenating contents column-wise
}
\examples{
d <- data.frame(a = c("Male", "0 - 20", 10),
                b = c("Male", "20 - 40", 8),
                c = c("Female", "40 - 60", 6))

collapse_row(d, 1:2)
}
