% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-row.R
\name{fill_row}
\alias{fill_row}
\title{Fill in missing values in rows with previous or next row value}
\usage{
fill_row(x, rows, direction = "right", browse = FALSE)
}
\arguments{
\item{x}{A dataframe}

\item{rows}{Rows to fill}

\item{direction}{Direction in which to fill missing values. Either "right" (the default) or "left.}

\item{browse}{debug}
}
\description{
This function is useful for tidying messy datasets, e.g. when importing data from formatted tables.
}
\examples{
d <- data.frame(a = c("Total", "0 - 20", 10),
                b = c(NA, "20 - 40", 8),
                c = c(NA, "40 - 60", 6))
fill_row(d, 1)
}
