% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-rows.R
\name{remove_rows_not_na_n}
\alias{remove_rows_not_na_n}
\alias{remove_rows_na_n}
\alias{remove_cols_na_n}
\alias{remove_rows_not_na_col}
\title{Remove Rows with limited Number of Non-Missing Values}
\usage{
remove_rows_not_na_n(x, n = 1)

remove_rows_na_n(x, n = 1)

remove_cols_na_n(x, n = 1)

remove_rows_not_na_col(x, col = 1, rows = NULL, .direction = "up")
}
\arguments{
\item{x}{a vector}

\item{n}{a integer vector. number of non-missing values}

\item{col}{a integer vector. columns to exclude when checking for NA rows}

\item{rows}{rows to check as a vector of row indices or as a regex}

\item{.direction}{in case that rows is a regex .direction sets the direction
in which to include the rows to check. One of "up" or "down"}
}
\description{
Remove Rows with limited Number of Non-Missing Values
}
\examples{
# ex-remove_rows_not_na_col
d <- data.frame(
  x = c(1, 2, 3, 5),
  y = c(1, 2, NA, NA),
  z = c(1, NA, NA, 3),
  w = c(1, NA, NA, 4)
)
# Remove rows which contain only NA in columns when excluding cols 1 and 2
remove_rows_not_na_col(d, col = c(1, 2))
# Remove all rows from the first two rows
# which contain only NA in columns when excluding cols 1 and 2
remove_rows_not_na_col(d, col = c(1, 2), rows = 1:2)
# Remove rows with exactly two NA
remove_rows_not_na_n(d, n = 2)
# Remove rows with exactly one or two NA
remove_rows_not_na_n(d, n = 1:2)
# Remove rows with exactly two non-NA values
remove_rows_na_n(d, n = 2)
}
