% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows2header.R
\name{rows2header}
\alias{rows2header}
\title{Make Column Header from rows}
\usage{
rows2header(data, rows, sep = "_", na.rm = TRUE, all.cols = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{rows}{rows to make the header from}

\item{sep}{seperator. Defaults to "_".}

\item{na.rm}{a logical. Remove NAs}

\item{all.cols}{Should all columns be renamed?}
}
\description{
Make Column Header from rows
}
\examples{
d <- data.frame(
  a = c("Total", "0 - 20", 10),
  b = c("Total", "20 - 40", 8),
  c = c("Total", "40 - 60", 6)
)
# Rename all columns
rows2header(d, 1:2)
# Do not rename column "b"
rows2header(d, 1:2, all.cols = "b")
# Do not rename columns 2:3
rows2header(d, 1:2, all.cols = 2:3)
# Rename only columns starting with `...`
rows2header(d, 1:2, all.cols = FALSE)
}
