% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows-to-names.R
\name{rows_to_names}
\alias{rows_to_names}
\alias{rows2header}
\title{Make Names from rows}
\usage{
rows_to_names(
  x,
  rows,
  sep = "_",
  na.sep = sep,
  na.rm = TRUE,
  all.cols = TRUE,
  regex = "^(\\\\.{3}|X\\\\d+)"
)

rows2header(
  x,
  rows,
  sep = "_",
  na.sep = sep,
  na.rm = TRUE,
  all.cols = TRUE,
  regex = "^(\\\\.{3}|X\\\\d+)"
)
}
\arguments{
\item{x}{A data frame}

\item{rows}{rows to make the header from}

\item{sep}{seperator. Defaults to "_".}

\item{na.sep}{seperator to be used for NAs if \code{na.rm=FALSE}. Defaults to \code{sep}.}

\item{na.rm}{a logical. Remove NAs}

\item{all.cols}{Should all columns be renamed?}

\item{regex}{A regular expression to select columns to rename if \code{all.cols = FALSE}.}
}
\description{
Make Names from rows
}
\examples{
d <- data.frame(
  a = c("Total", "0 - 20", 10),
  b = c("Total", "20 - 40", 8),
  c = c("Total", "40 - 60", 6)
)
# Rename all columns
rows_to_names(d, 1:2)
# Do not rename column "b"
rows_to_names(d, 1:2, all.cols = "b")
# Do not rename columns 2:3
rows_to_names(d, 1:2, all.cols = 2:3)
# Rename only columns starting with `...`
rows_to_names(d, 1:2, all.cols = FALSE)
}
