% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblcleanr-package.R
\docType{package}
\name{tblcleanr-package}
\alias{tblcleanr}
\alias{tblcleanr-package}
\title{tblcleanr: Cleaning data tables}
\description{
Utility functions to make it easy to clean real world data tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/trekonom/tblcleanr}
  \item Report bugs at \url{https://github.com/trekonom/tblcleanr/issues}
}

}
\author{
\strong{Maintainer}: Stefan Moog \email{moogs@gmx.de}

}
\keyword{internal}
