% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-remove-row.R
\name{str_remove_row}
\alias{str_remove_row}
\title{Remove matched patterns in rows}
\usage{
str_remove_row(x, rows, pattern)
}
\arguments{
\item{x}{dataframe}

\item{rows}{Rows to look for pattern.}

\item{pattern}{Pattern to look for.}
}
\description{
Remove matched patterns in rows
}
