% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-replace-row.R
\name{str_replace_row}
\alias{str_replace_row}
\alias{str_replace_row_na}
\title{Replace matched patterns in rows}
\usage{
str_replace_row(x, rows, pattern, replacement)

str_replace_row_na(x, rows, pattern)
}
\arguments{
\item{x}{dataframe}

\item{rows}{Rows to look for pattern.}

\item{pattern}{Pattern to look for.}

\item{replacement}{A character vector of replacements.}
}
\description{
Replace matched patterns in rows
}
\examples{
# TO DO
}
